Login adjustment for Web Version

+++ lib/screens/login_screen.dart
@@ -15,6 +15,7 @@
 *    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:picos/themes/global_theme.dart';
 import 'package:picos/util/backend.dart';
@@ -70,18 +71,24 @@ class _LoginScreenState extends State<LoginScreen>
       final String route = await Backend.getRole();
 
       if (_isChecked) {
-        await _secureStorage.setUsername(_loginController.text);
-        await _secureStorage.setPassword(_passwordController.text);
+        if (!kIsWeb) {
+          await _secureStorage.setUsername(_loginController.text);
+          await _secureStorage.setPassword(_passwordController.text);
+        }
       } else {
-        await _secureStorage.deleteAll();
+        if (!kIsWeb) {
+          await _secureStorage.deleteAll();
+        }
       }
       if (!mounted) return;
       Navigator.of(context).pushReplacementNamed(route);
     } else {
-      setState(() {
-        _backendError = loginError;
-        _sendDisabled = false;
-      });
+      setState(
+        () {
+          _backendError = loginError;
+          _sendDisabled = false;
+        },
+      );
     }
   }
 
@@ -89,7 +96,9 @@ class _LoginScreenState extends State<LoginScreen>
   void initState() {
     super.initState();
     Backend();
-    _fetchSecureStorageData();
+    if (!kIsWeb) {
+      _fetchSecureStorageData();
+    }
   }
 
   @override
@@ -177,21 +186,29 @@ class _LoginScreenState extends State<LoginScreen>
                 width: 200,
                 child: Row(
                   children: <Widget>[
-                    Text(AppLocalizations.of(context)!.rememberMe),
+                    !kIsWeb
+                        ? Text(AppLocalizations.of(context)!.rememberMe)
+                        : const SizedBox(
+                            width: 0,
+                          ),
                     Expanded(
-                      child: Checkbox(
-                        value: _isChecked,
-                        checkColor: Colors.white,
-                        activeColor: theme.darkGreen1,
-                        onChanged: (bool? value) {
-                          setState(
-                            () {
-                              _isChecked = value!;
-                              _secureStorage.setIsChecked(_isChecked);
-                            },
-                          );
-                        },
-                      ),
+                      child: !kIsWeb
+                          ? Checkbox(
+                              value: _isChecked,
+                              checkColor: Colors.white,
+                              activeColor: theme.darkGreen1,
+                              onChanged: (bool? value) {
+                                setState(
+                                  () {
+                                    _isChecked = value!;
+                                    _secureStorage.setIsChecked(_isChecked);
+                                  },
+                                );
+                              },
+                            )
+                          : const SizedBox(
+                              width: 0,
+                            ),
                     ),
                   ],
                 ),
@@ -201,8 +218,7 @@ class _LoginScreenState extends State<LoginScreen>
                 child: PicosInkWellButton(
                   disabled: _sendDisabled,
                   onTap: () {
-                    _submitHandler(
-                    );
+                    _submitHandler();
                   },
                   text: AppLocalizations.of(context)!.submit,
                 ),
